[REF] *: update typer and restructured code
Updated Typer to version 0.15.1, allowing for a cleaner separation of
all commands in subfolders etc. We now have one command per file and the
commands are grouped per scope in a subfolder.

Meanwhile we revised all code in order to make it clearer and shorter.

Updated the README to reflect the command name changes.