[FIX] po: don't overwrite with new po files
When executing the command `otk po create`, it would overwrite any
existing .po files with a new empty one, wiping out all translations.

Now, if there is already a .po file like the one we want to create, we
just leave it like that and display that the file already existed.