[IMP] dev: docker container start improvements
When starting the Docker containers, we will now check a config file to
see if the current version is different from the version of the last
time containers were launched. If so, we'll first stop the containers to
make sure all volume paths are updated.

When starting the dev server for the first time, we now indicate
separately the process of the Docker image being pulled.